pyproject.toml
[tool.poetry]
name = "book-club-training"
version = "0.1.0"
description = "Studying books with the book club of Alexey Goloburdin"
authors = ["ok_kir <okunev.kirill.ok@yandex.ru>"]
readme = "README.md"
packages = [{include = "book_club_training"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
aiohttp = "^3.8.4"
asyncpg = "^0.27.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
